Alterado tipo da funcao

+++ ILP-010/Listas/2-Funcoes recursivas/c.c
@@ -1,4 +1,4 @@
-int potencia(float b, int n) {
+float potencia(float b, int n) {
   // Receber um numero real b (b > 0) e um numero inteiro n (n >= 0) e devolver
   // o resultado de b elevado a n.
   if (n == 0)
